03_研究資料/research/analysis/requirements.txt
# 統合画像分類システム - 依存関係
# Python 3.8+ 推奨

# 基本的な科学計算・データ処理
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.0
Pillow>=8.3.0

# 機械学習・深層学習フレームワーク
torch>=1.11.0
torchvision>=0.12.0

# Transformers（BLIP, CLIP等）
transformers>=4.20.0
accelerate>=0.12.0

# CLIP（OpenAI）
git+https://github.com/openai/CLIP.git

# YOLO（Ultralytics）
ultralytics>=8.0.0

# SAM（Segment Anything）
git+https://github.com/facebookresearch/segment-anything.git

# 自然言語処理
nltk>=3.7.0

# GUI
tkinter  # 通常はPythonに標準搭載

# その他のユーティリティ
scikit-learn>=1.0.0
tqdm>=4.62.0
requests>=2.26.0

# オプション（環境に応じて）
# CUDA対応版のPyTorchを使用する場合：
# torch>=1.11.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.12.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html